requirements.txt
flask
requests
openai
line-bot-sdk
python-dotenv